MAINT: remove workflow submodule
