+++ actions/server.list.yaml
@@ -6,10 +6,6 @@ name: server.list
 parameters:
   timeout:
     default: 5400
-  submodule_name:
-    default: query
-    immutable: true
-    type: string
   action_name:
     default: list_all_openstack
     immutable: true
