+++ actions/server.search.by.datetime.yaml
@@ -6,10 +6,6 @@ name: server.search.by.datetime
 parameters:
   timeout:
     default: 5400
-  submodule_name:
-    default: query
-    immutable: true
-    type: string
   action_name:
     default: search_by_datetime
     immutable: true
