+++ actions/server.search.by.regex.yaml
@@ -6,10 +6,6 @@ name: server.search.by.regex
 parameters:
   timeout:
     default: 5400
-  submodule_name:
-    default: query
-    immutable: true
-    type: string
   action_name:
     default: search_by_regex
     immutable: true
