+++ tests/lib/workflows/test_search_by_property.py
@@ -2,19 +2,19 @@ from unittest.mock import patch, NonCallableMock, MagicMock
 import pytest
 
 from enums.query.sort_order import SortOrder
-from workflows.query.search_by_datetime import search_by_datetime
+from workflows.search_by_property import search_by_property
 
 
-@patch("workflows.query.search_by_datetime.openstack_query")
-@patch("workflows.query.search_by_datetime.QueryPresetsDateTime")
+@patch("workflows.query.search_by_property.openstack_query")
+@patch("workflows.query.search_by_property.QueryPresetsGeneric")
 @pytest.mark.parametrize(
     "output_type", ["to_html", "to_string", "to_objects", "to_props"]
 )
-def test_search_by_datetime_minimal(
+def test_search_by_property_minimal(
     mock_preset_enum, mock_openstack_query, output_type
 ):
     """
-    Runs search_by_datetime only providing required values
+    Runs search_by_property only providing required values
     """
 
     mock_query = MagicMock()
@@ -27,20 +27,15 @@ def test_search_by_datetime_minimal(
         "search_mode": NonCallableMock(),
         "property_to_search_by": NonCallableMock(),
         "output_type": output_type,
-        "days": 1,
+        "values": ["val1", "val2"],
     }
-    res = search_by_datetime(**params)
+    res = search_by_property(**params)
 
     mock_query.select_all.assert_called_once()
     mock_query.where.assert_called_once_with(
         preset=mock_preset_enum.from_string.return_value,
         prop=params["property_to_search_by"],
-        args={
-            "days": 1,
-            "hours": 0,
-            "minutes": 0,
-            "seconds": 0,
-        },
+        args={"values": params["values"]},
     )
     mock_query.run.assert_called_once_with(mock_cloud_account)
 
@@ -55,28 +50,29 @@ def test_search_by_datetime_minimal(
     )
 
 
-def test_search_by_datetime_errors_when_args_all_zero():
+def test_search_by_property_errors_when_no_values_given():
     """
-    test that search_by_datetime errors when days, hours, minutes and seconds are all set to zero
+    test that search_by_property errors when not provided a value to match against
     """
     with pytest.raises(RuntimeError):
-        search_by_datetime(
+        search_by_property(
             cloud_account=NonCallableMock(),
             query_type=NonCallableMock(),
             search_mode=NonCallableMock(),
             property_to_search_by=NonCallableMock(),
             output_type=NonCallableMock(),
+            values=[],
         )
 
 
-@patch("workflows.query.search_by_datetime.openstack_query")
-@patch("workflows.query.search_by_datetime.QueryPresetsDateTime")
+@patch("workflows.query.search_by_property.openstack_query")
+@patch("workflows.query.search_by_property.QueryPresetsGeneric")
 @pytest.mark.parametrize(
     "output_type", ["to_html", "to_string", "to_objects", "to_props"]
 )
-def test_search_by_datetime_all(mock_preset_enum, mock_openstack_query, output_type):
+def test_search_by_property_all(mock_preset_enum, mock_openstack_query, output_type):
     """
-    Runs search_by_datetime providing all available params
+    Runs search_by_property providing all values
     """
 
     mock_query = MagicMock()
@@ -90,24 +86,19 @@ def test_search_by_datetime_all(mock_preset_enum, mock_openstack_query, output_t
         "property_to_search_by": NonCallableMock(),
         "output_type": output_type,
         "properties_to_select": ["prop1", "prop2"],
-        "days": 1,
+        "values": ["val1", "val2"],
         "group_by": NonCallableMock(),
         "sort_by": ["prop1", "prop2"],
         "arg1": "val1",
         "arg2": "val2",
     }
-    res = search_by_datetime(**params)
+    res = search_by_property(**params)
 
     mock_query.select.assert_called_once_with(*params["properties_to_select"])
     mock_query.where.assert_called_once_with(
         preset=mock_preset_enum.from_string.return_value,
         prop=params["property_to_search_by"],
-        args={
-            "days": 1,
-            "hours": 0,
-            "minutes": 0,
-            "seconds": 0,
-        },
+        args={"values": params["values"]},
     )
     mock_query.sort_by.assert_called_once_with(
         *[(p, SortOrder.DESC) for p in params["sort_by"]]
