+++ tests/lib/workflows/test_search_by_regex.py
@@ -1,20 +1,18 @@
 from unittest.mock import patch, NonCallableMock, MagicMock
 import pytest
 
+from enums.query.query_presets import QueryPresetsString
 from enums.query.sort_order import SortOrder
-from workflows.query.search_by_property import search_by_property
+from workflows.search_by_regex import search_by_regex
 
 
-@patch("workflows.query.search_by_property.openstack_query")
-@patch("workflows.query.search_by_property.QueryPresetsGeneric")
+@patch("workflows.query.search_by_regex.openstack_query")
 @pytest.mark.parametrize(
     "output_type", ["to_html", "to_string", "to_objects", "to_props"]
 )
-def test_search_by_property_minimal(
-    mock_preset_enum, mock_openstack_query, output_type
-):
+def test_search_by_regex_minimal(mock_openstack_query, output_type):
     """
-    Runs search_by_property only providing required values
+    Runs search_by_regex only providing required values
     """
 
     mock_query = MagicMock()
@@ -24,18 +22,17 @@ def test_search_by_property_minimal(
     params = {
         "cloud_account": mock_cloud_account,
         "query_type": "MockQuery",
-        "search_mode": NonCallableMock(),
         "property_to_search_by": NonCallableMock(),
         "output_type": output_type,
-        "values": ["val1", "val2"],
+        "pattern": NonCallableMock(),
     }
-    res = search_by_property(**params)
+    res = search_by_regex(**params)
 
     mock_query.select_all.assert_called_once()
     mock_query.where.assert_called_once_with(
-        preset=mock_preset_enum.from_string.return_value,
+        preset=QueryPresetsString.MATCHES_REGEX,
         prop=params["property_to_search_by"],
-        args={"values": params["values"]},
+        args={"value": params["pattern"]},
     )
     mock_query.run.assert_called_once_with(mock_cloud_account)
 
@@ -50,29 +47,13 @@ def test_search_by_property_minimal(
     )
 
 
-def test_search_by_property_errors_when_no_values_given():
-    """
-    test that search_by_property errors when not provided a value to match against
-    """
-    with pytest.raises(RuntimeError):
-        search_by_property(
-            cloud_account=NonCallableMock(),
-            query_type=NonCallableMock(),
-            search_mode=NonCallableMock(),
-            property_to_search_by=NonCallableMock(),
-            output_type=NonCallableMock(),
-            values=[],
-        )
-
-
-@patch("workflows.query.search_by_property.openstack_query")
-@patch("workflows.query.search_by_property.QueryPresetsGeneric")
+@patch("workflows.query.search_by_regex.openstack_query")
 @pytest.mark.parametrize(
     "output_type", ["to_html", "to_string", "to_objects", "to_props"]
 )
-def test_search_by_property_all(mock_preset_enum, mock_openstack_query, output_type):
+def test_search_by_regex_all(mock_openstack_query, output_type):
     """
-    Runs search_by_property providing all values
+    Runs search_by_regex providing all values
     """
 
     mock_query = MagicMock()
@@ -82,23 +63,22 @@ def test_search_by_property_all(mock_preset_enum, mock_openstack_query, output_t
     params = {
         "cloud_account": mock_cloud_account,
         "query_type": "MockQuery",
-        "search_mode": NonCallableMock(),
         "property_to_search_by": NonCallableMock(),
         "output_type": output_type,
         "properties_to_select": ["prop1", "prop2"],
-        "values": ["val1", "val2"],
+        "pattern": NonCallableMock(),
         "group_by": NonCallableMock(),
         "sort_by": ["prop1", "prop2"],
         "arg1": "val1",
         "arg2": "val2",
     }
-    res = search_by_property(**params)
+    res = search_by_regex(**params)
 
     mock_query.select.assert_called_once_with(*params["properties_to_select"])
     mock_query.where.assert_called_once_with(
-        preset=mock_preset_enum.from_string.return_value,
+        preset=QueryPresetsString.MATCHES_REGEX,
         prop=params["property_to_search_by"],
-        args={"values": params["values"]},
+        args={"value": params["pattern"]},
     )
     mock_query.sort_by.assert_called_once_with(
         *[(p, SortOrder.DESC) for p in params["sort_by"]]
